First iteration of extracting the Order Details view data source to a sepatate class
At the moment, the view model is exposing the data source as a property, the viwemodel should probably implement the dataSource itself. Or if it exposes it as a property, the the table view could consume retain it and consume it directly
